requirements.txt
PySide6==6.9.2
PyYAML==6.0.2
opencv-python==4.9.0.80
Pillow>=10.0.0
imagehash==4.3.1
tqdm==4.66.4
